Calculate the 3db bandwidth of the filters
I noticed that on high frequencies the filters don't cover the frequency
spectrum.

Since the frequency resolution of the filters for high
frequencies is so fine, and since we have a fixed amount of filters for
each octave, this becomes a problem and we are blind to some
high frequencies that fall between bins. At the top octave, these gaps
are even 300% of the size of the neighboring filter's bandwidth!

We add error logs for this and hope to find a solution later.