Add average tuning drift statistic
This is visible in the debug view for now.

This is calculated as the absolute drift of each note from the nearest
semitone. This means that even music that is on average on pitch, i. e.
because of vibrato, will have a positive drift. But averaging over the
non-absolute drifts of the notes would be worse, because something at
the edge of +-50 cts would be displayed as perfect, and most randomly
drifting choirs would also average near "zero drift". The stricter
absolute pitch inaccuracy measure is probably more useful.